Moved functions from interactivepplanes to heterogenous_demos and refactor heterogenous NB
